ajout de commmentaires et readme
